refactor: move recalc_pointers to data_handling
